consolidate test configuration
enforce test isolation
