+++ Makefile
@@ -17,9 +17,9 @@ clean:
 	rm -rf include lib .Python bin
 
 tests:
-	bin/py.test micropsi_core micropsi_server --engine=dict_engine
+	bin/py.test --engine=dict_engine
 
 test-coverage:
-	bin/py.test micropsi_core micropsi_server --cov micropsi_core --cov micropsi_server --cov-report html
+	bin/py.test --cov micropsi_core --cov micropsi_server --cov-report html
 
 .PHONY: run
