+++ pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+norecursedirs=lib include dist spock src
